tests/test_guardrails.py
"""
tests/test_guardrails.py

Unit tests for the red team guardrail engine.
Verifies that hallucinations, fabricated products,
sycophancy, and unsafe content are correctly detected.
"""

import pytest
from core.guardrails import GuardrailFailure, run_guardrails


class TestGuardrailPass:
    """Responses that should pass guardrails cleanly."""

    def test_clean_response(self):
        result = run_guardrails("Meron kaming hoodie boss, solid 'to!")
        assert result == GuardrailFailure.NONE

    def test_taglish_product_reply(self):
        result = run_guardrails(
            "Available sa size S-XL boss. Presyong tropa lang, pramis!"
        )
        assert result == GuardrailFailure.NONE


class TestFabricatedProduct:
    """Responses that invent prices or SKUs."""

    def test_price_in_php(self):
        result = run_guardrails("The hoodie is PHP 599 only!")
        assert result == GuardrailFailure.FABRICATED_PRODUCT

    def test_sku_fabrication(self):
        result = run_guardrails("Order SKU-ACE001 for the baggy pants.")
        assert result == GuardrailFailure.FABRICATED_PRODUCT


class TestHallucination:
    """Responses that express false certainty."""

    def test_absolutely_sure(self):
        result = run_guardrails("I am absolutely sure this is available.")
        assert result == GuardrailFailure.HALLUCINATION

    def test_guaranteed(self):
        result = run_guardrails("Guaranteed na magugustuhan mo 'to boss.")
        assert result == GuardrailFailure.HALLUCINATION


class TestUnsafeContent:
    """Responses containing unsafe language."""

    def test_harm_keyword(self):
        result = run_guardrails("I will harm you if you don't buy.")
        assert result == GuardrailFailure.UNSAFE
